# Re-wrote mode functionality for better performance.  