--- a/util_test.go
+++ b/util_test.go
@@ -1,12 +1,23 @@
 package stats
 
 import (
+	"math/rand"
 	"testing"
 )
 
 // makeFloatSlice makes a slice of float64s
 func makeFloatSlice(c int) []float64 {
-	lf := []float64{}
+	lf := make([]float64, 0, c)
+	for i := 0; i < c; i++ {
+		f := float64(i * 100)
+		lf = append(lf, f)
+	}
+	return lf
+}
+
+func makeRandFloatSlice(c int) []float64 {
+	lf := make([]float64, 0, c)
+	rand.Seed(unixnano())
 	for i := 0; i < c; i++ {
 		f := float64(i * 100)
 		lf = append(lf, f)