Added tests, including generation of large random slices, to verify performance gains.

On my machine, the performance gain in benchmarks was 7x time reduction and 215x allocation reduction.