@Kunde21 2015-10-21
Commit hook for continuous-integration/wercker build has no example wercker.yml file in the repository.  I don't have the experience with wercker to blaze that trail, so I won't stumble through your library in order to learn.