@montanaflynn 2015-10-21
Thanks, always love to see performance improvements! 

I forgot to remove the wercker webhook after switching to travis CI, I just took care of that for future PR's!