@montanaflynn 2015-10-21
Benchmarks before and after (created with [benchcmp](https://godoc.org/github.com/polaris1119/go.tools/cmd/benchcmp)) for those interested:

```
benchmark nanosecond per op          old ns/op     new ns/op     delta
BenchmarkModeSmallFloatSlice         9973          3490          -65.01%
BenchmarkModeSmallRandFloatSlice     7569          2863          -62.17%
BenchmarkModeLargeFloatSlice         140343655     13242488      -90.56%
BenchmarkModeLargeRandFloatSlice     138829809     13302864      -90.42%

benchmark memory allocations         old allocs     new allocs     delta
BenchmarkModeSmallFloatSlice         10             4              -60.00%
BenchmarkModeSmallRandFloatSlice     6              3              -50.00%
BenchmarkModeLargeFloatSlice         6154           27             -99.56%
BenchmarkModeLargeRandFloatSlice     6152           27             -99.56%

benchmark bytes allocated            old bytes     new bytes     delta
BenchmarkModeSmallFloatSlice         432           208           -51.85%
BenchmarkModeSmallRandFloatSlice     312           160           -48.72%
BenchmarkModeLargeFloatSlice         10258315      4626432       -54.90%
BenchmarkModeLargeRandFloatSlice     10258084      4626432       -54.90%
```